chore: configure vault url in comparater server
